Enable sending metadata in ASGI responses
