--- sonora/context.py
@@ -1,77 +1,0 @@
-import time
-
-import grpc
-
-
-class gRPCContext(grpc.ServicerContext):
-    def __init__(self, timeout=None, metadata=None):
-        self.code = grpc.StatusCode.OK
-        self.details = None
-
-        self._timeout = timeout
-
-        if timeout is not None:
-            self._deadline = time.monotonic() + timeout
-        else:
-            self._deadline = None
-
-        self._metadata = metadata or tuple()
-
-    def set_code(self, code):
-        self.code = code
-
-    def set_details(self, details):
-        self.details = details
-
-    def abort(self, code, details):
-        if code == grpc.StatusCode.OK:
-            raise ValueError()
-
-        self.set_code(code)
-        self.set_details(details)
-
-        raise grpc.RpcError()
-
-    def abort_with_status(self, status):
-        if status == grpc.StatusCode.OK:
-            raise ValueError()
-
-        self.set_code(status)
-
-        raise grpc.RpcError()
-
-    def time_remaining(self):
-        if self._deadline is not None:
-            return max(self._deadline - time.monotonic(), 0)
-        else:
-            return None
-
-    def invocation_metadata(self):
-        return self._metadata
-
-    def send_initial_metadata(self, initial_metadata):
-        raise NotImplementedError()
-
-    def set_trailing_metadata(self, trailing_metadata):
-        raise NotImplementedError()
-
-    def peer(self):
-        raise NotImplementedError()
-
-    def peer_identities(self):
-        raise NotImplementedError()
-
-    def peer_identity_key(self):
-        raise NotImplementedError()
-
-    def auth_context(self):
-        raise NotImplementedError()
-
-    def add_callback(self):
-        raise NotImplementedError()
-
-    def cancel(self):
-        raise NotImplementedError()
-
-    def is_active(self):
-        raise NotImplementedError()
